+++ resources/js/app.js
@@ -7,16 +7,11 @@ import {
 } from "@inertiajs/inertia-vue";
 import { Link } from "@inertiajs/inertia-vue";
 import vuetify from "./plugins/vuetify";
-// import DatetimePicker from 'vuetify-datetime-picker'
-// import VueCtkDateTimePicker from 'vue-ctk-date-time-picker';
-// import 'vue-ctk-date-time-picker/dist/vue-ctk-date-time-picker.css';
 
 
 import { InertiaProgress } from "@inertiajs/progress";
 
-//Vue.component('VueCtkDateTimePicker', VueCtkDateTimePicker);
 Vue.use(InertiaPlugin);
-//Vue.use(DatetimePicker);
 
 Vue.component("Link", Link);
 Vue.mixin({ methods: { route: window.route } });
