+++ resources/js/plugins/vuetify.js
@@ -16,7 +16,7 @@ const options = {
         info: colors.teal.lighten1,
         warning: colors.amber.base,
         error: colors.deepOrange.accent4,
-        success: colors.green.accent3,
+        success: colors.green.accent4,
         dirneder: colors.green.darken2
       }
     }
