+++ resources/lang/de/passwords.php
@@ -13,7 +13,7 @@
 
 return [
     'reset'     => 'Das Passwort wurde zurückgesetzt!',
-    'sent'      => 'Passworterinnerung wurde gesendet!',
+    'sent'      => 'Mail zur Passworteränderung wurde gesendet!',
     'throttled' => 'Bitte warten Sie, bevor Sie es erneut versuchen.',
     'token'     => 'Der Passwort-Wiederherstellungs-Schlüssel ist ungültig oder abgelaufen.',
     'user'      => 'Es konnte leider kein Nutzer mit dieser E-Mail-Adresse gefunden werden.',
